Updated openstack/openstack Project: openstack/openstack-manuals  d428cd0f4c56073be8c9f56acda4a08fb6022743
Add doc for Clone CG

In Liberty, we introduced support for creating a
consistency group from a source consistency group.
This patch adds document for it. This patch also
adds a missing entry in the policy file for update
CG.

Cinder server side change is merged:
https://review.openstack.org/#/c/203473/

Cinder client side change is merged:
https://review.openstack.org/#/c/205897/

Change-Id: I4e9a2b42f1a55c2351c9b78259a71593e0ac21e5