Updated openstack/openstack Project: openstack/heat  1397af1b6d0ea3591c65055ae318b663b0e92d9f
Convergence: Fix failing integration tests

Input data can contain tuple as key when the attribute and path
components are resolved. Converting this to JSON (serializing) fails.
To fix this, recursively look for tuple as keys in input data and
convert them to string and vice-versa while serializing and
deserializing.

Change-Id: I87e496d51004f3374965332921628f5eccb34657
Partial-Bug: #1492116