Updated openstack/openstack Project: openstack/zaqar  7aae02d3c74bb54bb9b4d17d29e95e42bb08918a
Use a Text field instead of Binary

Some database engines require for Binary fields to have a size. For now,
instead of using a Binary field and guess the size, we should just use a
Text field.

Some engines may require a lenght for this field but that's not the case
for the majority we're interested in supporting for now.

Change-Id: I06883c9dd78b2d6ab83217d2d171298891483e9f