Updated openstack/openstack Project: openstack-infra/project-config  05ac1f2faaebb7534d45adb85b1ee5fa9b9055ef
Make gate-tempest-dsvm-ironic-pxe_ssh vote again

This is no longer running against Nova. It *is* still talking to Ubuntu
servers. There is a risk of this breaking due to things Ubuntu does out
of our control; however it should now only break the Ironic gate.

The DIB-built ramdisk for this job is deprecated and being removed at
the beginning of M, and the job will be removed with it. Until then, I
think it's a reasonable risk to take to ensure we don't break
compatibility with this ramdisk.

This also removes the pxe_ssh job from ironic's gate, so that if it does
fail it won't be resetting other project gates all over the place.

Change-Id: I00c9c93f0e624ca6779c8334e2f6f2a19efe34e1