Updated openstack/openstack Project: openstack-infra/project-config  767b508d8203e2b0de4d812ea9c0fb21ad2467ab
Add announcements for keystoneauth to #openstack-sdks

Since keystoneauth will be very tightly coupled to the SDK efforts,
ensure that proposals and merges are announced to the #openstck-sdks
channel. This is being added after a brief discussion with Terry
Howe (who is a core on keystoneauth as well as heavily involved in
the SDK efforts).

Change-Id: I62c83d7598ca3661c5d3a4a21a968914f53a46cf