Updated openstack/openstack Project: openstack/sahara  b6338ecfd85a32951a243d4679feaefa01e82b58
Only add current directory to classpath for client deploy mode

We added a ":" to the spark driver-classpath to fix spark/swift
integration for spark 1.3.1. However, the ":" is only necessary
for client deploy mode so we should leave it out in other cases.

Note, this is a refinement to the original bug fix.  The ":" in
other deployment modes does not break anything.  However, it is
better not to include it since it is unnecessary.

Partial-bug: #1486544
Change-Id: Iaacbb090d0065922fab034d9d4a1f765ad7e05e3