Updated openstack/openstack Project: openstack-infra/project-config  04cbc5a902159258cbb683e1284e26ecca74ce41
Make gate-horizon-python34 voting and add gate

The py34 gate was fixed, it runs a subset of tests on Python 3.4.
It's now important to make the py34 gate voting to avoid Python 3
regressions.

The Horizon port to Python 3 is not complete, more code will be
ported slowly. We adopted the same strategy in other projects: nova,
cinder, glance, etc.

Partial-Implements: blueprint porting-python3
Change-Id: I796c838f134f677f0f840c33e689bf713def1202