Updated openstack/openstack Project: openstack/nova  e1848f33971804806df8731f654de46a57185675
Add API schema for different_cell filter

For using the different cell filter, the API parameter different_cell
needs to be passed to the corresponding filter. However v2.1 API
validation denies it because there is not API parameter definition
and it didn't work at all.
This patch adds the API schema.

Change-Id: Ic61b9a18722b7701b09e71f3f3a547ad9e6363c2
Closes-Bug: #1492925