Updated openstack/openstack Project: openstack/ceilometer  72eb6985c47e87c77e61c8a36b258f40b0f87c47
Add attribute 'state' to meter metadata when source is polling

For now, instance attribute 'state' is stored to
metadata of meter which was generated from
notifications. But when meter was generated from
instance polling, 'state' is not stored to
metadata of meter even this attribute could be
retrieved from polling action. Because the meter
generated from notifications or polling update
the resource information in database in cycles,
the resource show API may return the instance
metadata with different style(sometimes the
returned metadata contains 'state', but at other
times, the 'state' is disappeared in it), which
may confuse us in some situations. In this patch,
'state' has been added to the meter metadata if
the meter was generated from polling source.

Closes-Bug: #1491224

Change-Id: Ie869672901b1d13f9444d0ca0e7d849f5e4ffc4b