Updated openstack/openstack Project: openstack/heat  e6f375e8fc3fc51ee9c923f067dc3b8409e02e18
Work around parameter error deleting backup stack

There is a problem where restoring a backup resource which references
a parameter in it's properties that doesn't exist in the backup stack
will fail to delete, if the property is referenced in the delete path
of the resource.

We can work around the problem by always returning the frozen_definition
properties on delete, provided there are stored properties to refer to.

Closes-Bug: #1494260
Change-Id: I462ce7161497306483286b78416f9037ac80d6fa