Updated openstack/openstack Project: openstack-infra/puppet-httpd  84c5d2bf6a20ce6c4ad8d900e4cd4a8ccfda143c
Fix httpd::dev.

While writing acceptance tests for puppet-httpd we noticed that the
manifest of httpd::dev was not working on Ubuntu Trusty.

That is because httpd::dev was trying to install multiple package using
the wrong syntax. Also, the one of the modules for Ubuntu was wrong and
we updated to keep applying.

Co-Authored-By: Bruno Tavares <btavare@thoughtworks.com>
Co-Authored-By: Danilo Ramalho <dramalho@thoughtworks.com>

Change-Id: Ibe1e107417b4670902f74b8c2fedc0d9d927ce9b