Updated openstack/openstack Project: openstack-infra/project-config  78ddeb7bcc437d68ffab4399716ef654dad3341b
Stop pushing to Transifex

This removes the code from all four upstream_translation_* scripts
that pushes new .pot files up to Transifex, and forces the comments
around them to be in line.

This leaves Transifex setup in place for the moment, since the
propose jobs that pull from Transifex are still utilising it.

Change-Id: I702e6073f92faaa79a7aa53c44022f9e437721af