Updated openstack/openstack Project: openstack-infra/project-config  6bdd5f8b862cd9e5b3d98d6e040a737b1f212821
Refactor verification for ZanataRestService

Rather than pass verify to each function that requires it, refactor
it down the stack into __init__. Now that we have moved to using
translate.o.o which has a trusted certificate, switch the default
to True.

Add a command-line argument --no-verify to create-zanata-xml to ease
testing against translate-dev.

Change-Id: I0f2314a8e82c3b7f1c96f28408c36ea1dd786e2f