Updated openstack/openstack Project: openstack/manila  c5c13f26f0d9c801196b1c190721ebe5990414e0
NetApp cDOT driver should prefer aggregate-local LIFs

The NetApp driver returns the export locations in the same order
regardless of the locations of the LIFs and aggrs. If the flexvol
containing the share is on a node with a LIF, that LIF should be
sorted to the top of the list of the export locations.

Change-Id: I36d6b04849899ee3378c8107381f29313d9e35f4
Closes-Bug: 1485097