Updated openstack/openstack Project: openstack/horizon  c55f0df912c7961ceafde9d8b3c873ff248ac77f
hz-page-header should allow string interpolations

As noted by Thai in his review for the patch
https://review.openstack.org/#/c/213879/, the hz-page-header
directive should be updated to allow string interpolations
such as {$ ... $} in order to support angular translate.

This patch can be tested by setting DISABLED = False in
_1051_project_ng_images_panel.py and _3031_identity_users_panel.py
and checking the header translation with the pseudo
translation tool.

Change-Id: If9778f361d2bb1682e258578b7525bb734b8f8b6
Closes-Bug: 1485778