Updated openstack/openstack Project: openstack/python-novaclient  fb47954802657ec441a4b7f659e42f6f30fa7302
Add response message when the state of a server is reset

When a request is made for reset the server state it will not print any
success message if it got reset successfully. The error message is
already getting printed when the reset action fails. This fix will
print response message if the reset action is success.

Request>> nova reset-state --active <UUID>
Output Message will print as: Reset state for server <UUID> succeeded;
new state is active

    Closes-Bug: #1480428

Change-Id: Id069bc55fd88ce7e62f82f20e47be5373bb6ce8f