Updated openstack/openstack Project: openstack/python-novaclient  e1346865db24dd1d211446eb8508b052e21ff39b
Modify "nova keypair-show" Positional arguments help information

CLI "nova keypair-show <keypair>", Positional arguments:
  <keypair> Name or ID of keypair
But in fact, it does not support ID,when using ID, the error is as follows:
[root]# nova keypair-show 631
ERROR (CommandError): No keypair with a name or ID of '631' exists.
So it is needed to change the Positional arguments help information
from "Name or ID of keypair" to "Name of keypair".

Change-Id: I4cd022a6aa1ca937e7aeb33847dc3ce2d6d5690e
Closes-Bug: #1493212
Signed-off-by: Xiaowei Qian <qianxw.fnst@cn.fujitsu.com>