Updated openstack/openstack Project: openstack/nova  a58ab9dc8d2b9a8a21ffa31a5f91b081c49f7b37
libvirt-vif: Allow to configure a script on bridge interface

While running with libvirt-xen driver, it is possible to have the Xen
toolstack running a different script than the default on a vif. This patch
allow Nova to change this script.

Also do not set script to the empty string '' in designer.py for a linux
bridge. The empty string for script does not appear to be use anywhere in
the libvirt code when the vif is a bridge.

Change-Id: Ib6d6542d22decccfa68a058d362a42d60e6c2cca
Partial-Bug: #1461642