Updated openstack/openstack Project: openstack/ironic  3cafd272f8a94a54687b84b6e16993cd38bce1e2
Agent supports post-clean-step operations

This adds a new decorator in agent_base_vendor:
post_clean_step_hook(interface, step). A method with that decorator
will be invoked after the successful completion of the (clean) step.
The method itself should take two arguments: 1. the task, 2. the
command object corresponding to the completed clean step that was
returned from IPA as a result of a 'get_commands_status' request to
IPA.

Change-Id: I2d6ef6f81c134e9b3d38fea956e86cfc5c3c4ad4
Depends-On: I85220ffe9842576d73b91d194cfdaca6c437ebd2
Implements: blueprint inband-raid-configuration
Co-Authored-By: Ramakrishnan G <rameshg87@gmail.com>