Updated openstack/openstack Project: openstack/tempest-lib  faa6b9a902fb1aa2963e77b653a1a6fa1281f3e4
Write the latest change_id on migration commit

The previous script wrote a lot of change_id on the migration commit
message, and that was unreadable. Developers needed to change the
message by hands for showing the latest change_id for each migrated
module. That was boring.

This patch makes the script write the latest change_id like
 * run_tempest.sh: I55311cff4dea1f6c5adaa2a6d51eeee2f6cc71ea
 * run_tests.sh: I55311cff4dea1f6c5adaa2a6d51eeee2f6cc71ea

Change-Id: I7ce8e7b911c97ad7757a886f781341cce74c5307