Updated openstack/openstack Project: openstack/neutron  ca5fd3cd11f3b653bb91ff12964fb26b2b946b32
Prevent full sync in dhcp_agent when possible

If an exception occurs in sync_state method, we try for a full sync
even in the case where we have fewer networks to resync for.

This turns out to be pretty costly in scaled environments.

This patch addresses the above behavior by resyncing only for the
eariler set of failed networks.

Closes-Bug: #1495592

Change-Id: I069e992b3b7814370d409236b6a3c81a25829cc1