Updated openstack/openstack Project: openstack/trove  17ac1d0a17624ed641e1eba5b0246d17220e9a69
Fix get_flavors test by sorting json output

The get_flavors unit test sometimes fails as the list can
come back in a different order.  This causes the json comparison
to inacurately claim that the content is different.  Since
we don't care what order the json data is stored, this
comparison is now done on a sorted data set.

Closes-Bug: #1437179
Change-Id: I92eaa1c98f5a58ce124210f2b6a2136dfc573a29