Updated openstack/openstack Project: openstack/horizon  eedcdb4d50eae42d592b9fd4b164c2df9a5f9bd1
Fix "create failed" bug in add DHCP agent form

When Add DHCP agent for a network that it's name is blank,
it cast error "This field is required" beside the network
name field. So there should display the network's name_or_id
in "Add DHCP Agent" form modal, and this is consistent with
the "Edit Network". This patch fix the bug.

Change-Id: I4ad4b1d9907d7aa521ba6da44be8230aa301fcb2
Closes-Bug: #1495105