Updated openstack/openstack Project: openstack-infra/project-config  1987aab3135484439c3908a565c94a25a4e7a785
Drop lib-forward-testing for python-ironicclient

This job is currently failing, because it doesn't install ironicclient
from git properly. As much as I'd love to fix this, I don't think this
test is actually useful.

AIUI, the test runs full tempest. However, it does not use Nova's Ironic
driver, which is the only part of full tempest that would use
python-ironicclient. So I don't think any value is actually gained from
this test.

Change-Id: I7941c6eb3129ba41d263c53149be31072ac67cb0