Updated openstack/openstack Project: openstack-infra/elastic-recheck  66b8491a66b7b59cb05fe3c0a7a06fb1c6452c35
Add query for neutron constraints job bug 1496424

This is a non-voting job in the feature/pecan branch, so it's not
something that elastic-search will comment on in a change. However,
it's blowing up the uncategorized bugs page, so we can categorize
it here to cleanup that page.

Also add TODO for tracking voting flag in elasticsearch.

This comes up from time to time, so add a TODO that we need to track
whether or not a job is voting or not in elasticsearch, then we can use
that to filter out non-voting jobs from the uncategorized bugs page.

Related-Bug: #1496424

Change-Id: I820cce8e9a615732a5a81472fd40957db14ee2c9