Updated openstack/openstack Project: openstack-infra/subunit2sql  a8af87bd093bb4e37cf5850b9af45836fb9e47ca
Fix for sqlite support on migrations

SQLite support was broken by a migration which needed an ALTER TABLE
on a constraint of test_runs as ALTER TABLE support on SQLite is
very shallow.
Alembic developers suggest using batch_alter_table method for these
situations http://alembic.readthedocs.org/en/latest/batch.html.

Besides the batch_alter_table change a forced drop_index is needed
for sqlite provider as index are created on a global namespace and
it couldn't create a new temp table with an index with the same_
name.

Change-Id: I287bc710369325bc626b62aea7eb6d0f46f8314a