Updated openstack/openstack Project: openstack/python-keystoneclient  f7829dd7f64b85cbddc141e122c3370b33964696
Avoid message concatenation in error path

Recently, the error message in _process_communicate_handle_oserror()
has been i18n'ed, which caused the regression as another code path
appended a string to it, which causes the TypeError to be raised.

Fix it by using string formatting instead of '+' to force it
to convert to string before concatenating.

Closes-Bug: 1421652

Change-Id: I7229b46888f798ac4a69c140ab389afed49b8c3c