Updated openstack/openstack Project: openstack-infra/subunit2sql  580af9b48acbb61054f7cffb7a49fe0991ad29b3
Add support to more than one db provider for tests

As db api worked with one EngineFacade singleton, the migrations tests
could not run successfully to MySQL and PostgreSQL database. The
migrations started being applied to the MySQL database (and EngineFacade
was instanciated based on it) and after to PostgresSQL, but EngineFacade was
not instantiated anymore, keeping with the MySQL engine.

In order to fix that issue, this commit implement a way so that the db api
works with one EngineFacade by database vendor.

Co-Authored-By: Moises Trovo <mtrovo@thoughtworks.com>

Change-Id: I07898bd963c08e7a2eff946b8f0c0886f2143474