Updated openstack/openstack Project: openstack/taskflow  f8624a136b77c68dbf3e75406f2903586cf0762b
Fix '_cache_get' multiple keyword argument name overlap

The argument 'atom_name' is also used by the fetch function
so when it is provided a conflict occurs and this ends badly.

To avoid this capture the needed variables used for fetching
a cached value into a functools.partial object and use that instead
of passing further arguments.

Closes-Bug: #1496608

Change-Id: Ic012f7687037bf876d041c4bc62b3f6606a8a845