Updated openstack/openstack Project: openstack/horizon  d6ce0cf9f0fb809b241ab68118cb47f722e6501b
[Sahara] Unified job interface map UI

A UI to serve the "Unified Job Interface Map" feature in Sahara. This
feature adds an "interface" map to the API for job creation, such that
the operator registering a job can define a unified, human-readable way
to pass in all arguments, parameters, and configurations that the
execution of that job may require or accept. This will allow platform-
agnostic wizarding at the job execution phase and allows users to
document use of their own jobs once in a persistent, standardized format.

Change-Id: I81527405aa8038b13ff6e55ac5d972775ae810a6
Implements: blueprint unified-job-interface-map-ui