Updated openstack/openstack Project: openstack/horizon  ab87e0275612780b4b80207cd4f0a5ed8719a1bd
Added a django extract for .csv files

The CSV files from the overview panels were mostly not 
translated when using the pseudo translation tool. We 
inspected the django.pot files and the messages from the CSV 
files were not present.

This patch updates the babel-django config file to ensure that 
*.csv files are included for translation extraction.

We tried to update the babel-django config file with one 
django extraction line which included both the html file 
descriptor and the csv file descriptor separated by a comma 
but testing showed this did not work properly.  We could only 
get this to work with two separate extraction lines.

Change-Id: I04b3a1f49c7c680a4f3dfdc336183749404973ec
Closes-Bug: #1496080