Updated openstack/openstack Project: openstack/manila  5ededfa4dcc0e4cbd5257e98749bd7cd2f4e95c9
Remove IBM GPFS driver due to lack of CI

The Manila community set a deadline of 3 Sep 2015 for all drivers
to have reporting CI systems, and there is no CI system for the
IBM GPFS driver. As agreed, the driver will be removed until such
time that a CI system exists.

Change-Id: I76f238a1a6eb039e4ddc1cb189115ab3e686c56b