Updated openstack/openstack Project: openstack-infra/elastic-recheck  0ea6f96407883921a96a7d9167a00e953408cab7
Add signatire for 1496678

oslo.utils 1.4.1 was release today to unlock several stable/juno issues. Sadly
this version is also valid in stable/kilo.

The Babel entry in juno is too restrictive for kilo.

This is fixed already but we should categorize the bug so it's not in the
uncategorized bugs page (redundant, I know).

229 hits in 7 days, check and gate, all failures.

Change-Id: Iaed30a2a1f0a000caa39229bd053236d84991913
Related-Bug: 1496678