Updated openstack/openstack Project: openstack/heat  320ae5f684c12bc0bdd02c993bb161f509eb14f8
Allow map/list items for str_replace

Currently any attempt to substitute placeholders using data from a
json/map or list parameter/attribute results in an error, which is
inconvenient if you wish to provide such data to an instance, e.g
so it can be used in a script via jq/python or whatever.

So, tolerate non-string parameter values, and attempt to serialize
them as json, then substitute the resulting json strings.

Change-Id: I362cc76ac3f68d4649a62459455c0dcae2dcd25d
Closes-Bug: #1489028