Updated openstack/openstack Project: openstack/nova  d7de04240dbfbda854875bf878ee24e7c1b7c2d8
Xenapi: Don't access image_meta.id when booting from a volume

When a request is made to boot from a preexisting volume, rather than
asking Nova to create a volume from an image, there is no image id field
in the image_meta that's passed in to the virt driver.  This would be
fine except that there was an attempt to access image_meta.id in the
boot from volume case where it is not necessary.  This change moves that
access to a point where it is known that this is not an attempt to boot
from a volume and that field should be set.

Change-Id: I9dab14762f5d331ef2e36f617f7f66c3d771f32d
Closes-bug: 1496557